refactor: MovementStrategy 클래스명 Movement로 변경 및 패키지 이동
