+++ src/main/java/domain/movement/LinearMovement.java
@@ -1,12 +1,11 @@
-package domain.strategy;
+package domain.movement;
 
 import domain.Direction;
 import domain.Position;
-
 import java.util.ArrayList;
 import java.util.List;
 
-public class LinearMovement extends MovementStrategy {
+public class LinearMovement extends Movement {
 
     @Override
     protected List<Position> buildRoute(List<Direction> path, Position sourcePosition, Position targetPosition) {
