+++ src/main/java/domain/movement/Movement.java
@@ -1,10 +1,10 @@
-package domain.strategy;
+package domain.movement;
 
 import domain.Direction;
 import domain.Position;
 import java.util.List;
 
-public abstract class MovementStrategy {
+public abstract class Movement {
 
     protected final String INVALID_TARGET_POSITION = "이동할 수 없는 목적지입니다.";
