+++ src/main/java/domain/movement/PathMovement.java
@@ -1,11 +1,11 @@
-package domain.strategy;
+package domain.movement;
 
 import domain.Direction;
 import domain.Position;
 import java.util.ArrayList;
 import java.util.List;
 
-public class PathMovement extends MovementStrategy {
+public class PathMovement extends Movement {
 
     @Override
     protected List<Position> buildRoute(List<Direction> path, Position source, Position target) {
