+++ src/main/java/domain/piece/Elephant.java
@@ -3,7 +3,7 @@ package domain.piece;
 import domain.Direction;
 import domain.Position;
 import domain.Side;
-import domain.strategy.MovementStrategy;
+import domain.movement.Movement;
 import java.util.List;
 
 public class Elephant extends Piece {
@@ -19,13 +19,13 @@ public class Elephant extends Piece {
         List.of(Direction.LEFT, Direction.DOWN_LEFT, Direction.DOWN_LEFT)
     );
 
-    public Elephant(Side side, MovementStrategy movementStrategy) {
-        super(side, movementStrategy);
+    public Elephant(Side side, Movement movement) {
+        super(side, movement);
     }
 
     @Override
     public List<Position> findRoute(Position sourcePosition, Position targetPosition) {
-        return movementStrategy.findRoute(paths, sourcePosition, targetPosition);
+        return movement.findRoute(paths, sourcePosition, targetPosition);
     }
 
     @Override
