+++ src/test/java/domain/piece/PieceTest.java
@@ -2,7 +2,7 @@ package domain.piece;
 
 import domain.Position;
 import domain.Side;
-import domain.strategy.PathMovement;
+import domain.movement.PathMovement;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
